[feat]: Implement RTC driver

+++ Kernel/API/POSIX/sys/types.h
@@ -20,23 +20,23 @@ extern "C"
 #undef unsigned
 #pragma clang diagnostic pop
 
-    typedef int64_t  blkcnt_t;
-    typedef int64_t  blksize_t;
-    typedef uint64_t clock_t;
-    typedef uint64_t clockid_t;
-    typedef uint64_t dev_t;
-    typedef uint64_t fsblkcnt_t;
-    typedef uint64_t fsfilcnt_t;
-    typedef uint32_t gid_t;
-    typedef int32_t  id_t;
-    typedef uint64_t ino_t;
-    typedef uint64_t key_t;
-    typedef uint16_t mode_t;
-    typedef uint64_t nlink_t;
-    typedef int64_t  off_t;
-    typedef int32_t  pid_t;
-    typedef uint32_t uid_t;
-    typedef int64_t  time_t;
+    typedef int64_t      blkcnt_t;
+    typedef int64_t      blksize_t;
+    typedef uint64_t     clock_t;
+    typedef int          clockid_t;
+    typedef uint64_t     dev_t;
+    typedef uint64_t     fsblkcnt_t;
+    typedef uint64_t     fsfilcnt_t;
+    typedef uint32_t     gid_t;
+    typedef int32_t      id_t;
+    typedef uint64_t     ino_t;
+    typedef uint64_t     key_t;
+    typedef unsigned int mode_t;
+    typedef uint64_t     nlink_t;
+    typedef int64_t      off_t;
+    typedef int32_t      pid_t;
+    typedef uint32_t     uid_t;
+    typedef int64_t      time_t;
 
 #ifdef __cplusplus
 }
